% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrongler_period.R
\name{as.matrix.list}
\alias{as.matrix.list}
\title{Convert List of chrongler_period Objects to a Matrix}
\usage{
\method{as.matrix}{list}(x, ...)
}
\arguments{
\item{x}{A list of \code{chrongler_period} objects.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
A matrix if the list contains \code{chrongler_period} objects; otherwise, a fallback matrix.
}
\description{
This function converts a list of \code{chrongler_period} objects to a matrix. If all elements of
the list are \code{chrongler_period} objects, it coerces them to character vectors and binds them
into a matrix. Otherwise, it falls back to the default behavior of \code{as.matrix}.
}
