% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrongler_period.R
\name{as.vector.chrongler_period}
\alias{as.vector.chrongler_period}
\title{Coerce chrongler_period to a Vector}
\usage{
\method{as.vector}{chrongler_period}(x, ...)
}
\arguments{
\item{x}{A \code{chrongler_period} object.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
A character vector with the values from the \code{chrongler_period} object.
}
\description{
This function coerces a \code{chrongler_period} object into a character vector.
It is dispatched when \code{as.vector} is called on an object of
class \code{chrongler_period}.
}
