% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BuildingsMilet}
\alias{BuildingsMilet}
\title{Periods of Selected Buildings from the Map of Miletus}
\format{
\subsection{\code{BuildingsMilet}}{

A data frame with 110 rows and 3 columns:
\describe{
\item{identifier}{The name of a building in Miletus.}
\item{period.start, period.end}{The earliest and latest period each building is dated to.}
}
}
}
\source{
\url{https://geoserver.dainst.org/catalogue/#/map/5764}
}
\usage{
BuildingsMilet
}
\description{
A subset of data from the Map of Miletus for testing and demonstrating
working with the R-Package chrongler.
}
\keyword{datasets}
