% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_chrongler_conc.R
\name{chrongler_concordance}
\alias{chrongler_concordance}
\title{Produce a concordance object that can be used with \code{chrongler} functions}
\usage{
chrongler_concordance(x)
}
\arguments{
\item{x}{A list of chrongler_periods}
}
\value{
A list of S3-class chrongler.conc
}
\description{
Produce a concordance object that can be used with \code{chrongler} functions
}
\examples{
print("not now")
}
