% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_period.R
\name{derive_period}
\alias{derive_period}
\title{Derive Period from Absolute Dates}
\usage{
derive_period(data, conc, min, max, previous_start, previous_end)
}
\arguments{
\item{data}{A data.frame containing at least two columns with the
minimum and maximum absolute dating for each row as integer.
The values need to correspond with the possible dating ranges present in
the concordance used (see \code{\link[=make_chrongler_conc]{make_chrongler_conc()}}).}

\item{conc}{\emph{chrongler.conc}, \emph{required}, as built by \code{\link[=make_chrongler_conc]{make_chrongler_conc()}}.}

\item{min}{\emph{chr/int}. Name or index of the column with the \strong{minimum dating} (integers representing years).}

\item{max}{\emph{chr/int}. Name or index of the column with the \strong{maximum dating} (integers representing years).}

\item{previous_start}{\emph{chr/int}, \emph{optional}. Name or index of a pre-existing column
containing a \strong{starting period}. If supplied, previously empty values (\code{NA})
will be populated in the new \emph{period.start} column. A comment is stored in a
column called \emph{period.source}.}

\item{previous_end}{\emph{chr/int}, \emph{optional}. Name or index of a pre-existing column
containing a \strong{ending period}. If supplied, previously empty values (\code{NA})
will be populated in the new \emph{period.end} column. A comment is stored in a
column called \emph{period.source}.}
}
\value{
The input \code{data} as a \code{data.frame} with three additional columns:
\itemize{
\item \code{period.start} -- character, the period the date in \code{min} would fall in
(or copied from \code{previous_start} where available)
\item \code{period.end} -- character, the period the date in \code{max} would fall in
(or copied from \code{previous_end} where available)
\item \code{period.source} -- character, a comment indicating how the period was
derived (\code{"Derived from absolute dating"}, \code{"Partially derived from absolute dating"},
or \code{NA} if no derivation was necessary)
}
}
\description{
For each row in \code{data}, the value in the \code{min} and \code{max} column
is used to identify the period(s) the respective row would belong to
according to the dating ranges registered in the the concordance
object supplied to \code{conc} (\code{\link[=make_chrongler_conc]{make_chrongler_conc()}}). The newly assigned
periods are stored in column names \code{period.start} and \code{period.end}.
If \code{previous_start} and/or \code{previous_end} are supplied, the values from those
columns are preferred, and only empty (\code{NA}) cells are filled with derived periods.
A comment informing of the procedure is written to a column called
\emph{period.source}.
}
\examples{
conc <- make_chrongler_conc(RomanPeriods)

data <- data.frame(name = 1:110)

data$min <- sample(seq(-509, 476, by = 1), 110, replace = TRUE)
data$max <- data$min + sample(abs(round(rnorm(1000, 200, 100))), 110, replace = TRUE)

derive_period(data, conc, min = "min", max = "max")

}
\seealso{
\itemize{
\item \code{\link[=make_chrongler_conc]{make_chrongler_conc()}}
\item \code{\link[=derive_dating]{derive_dating()}}
}
}
