% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_chronontology.R
\name{dl_chrnt_json}
\alias{dl_chrnt_json}
\title{Download List from iDAI.chronOntology}
\source{
See docs at https://github.com/dainst/chronontology-backend/blob/master/docs/rest-api-reference.md
}
\usage{
dl_chrnt_json(id = "")
}
\arguments{
\item{id}{(chr) The chronOntology-ID of the period that should be
downloaded (part of the URL after
"https://chronontology.dainst.org/period/period/").}
}
\value{
A nested list formatted from JSON.
}
\description{
Download List from iDAI.chronOntology
}
\keyword{internal}
