% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrongler_period.R
\name{==.chrongler_period}
\alias{==.chrongler_period}
\title{Compare Two chrongler_period Objects}
\usage{
\method{==}{chrongler_period}(a, b)
}
\arguments{
\item{a}{A \code{chrongler_period} object.}

\item{b}{Another \code{chrongler_period} object.}
}
\value{
A logical value (\code{TRUE} or \code{FALSE}), indicating the result of the comparison.
}
\description{
These functions allow comparison between two \code{chrongler_period} objects based on their attributes.
}
\details{
The comparison operators for \code{chrongler_period} objects are defined as follows:
\itemize{
\item \code{==}: Compares the \code{name}, \code{start_date}, and \code{end_date} of two \code{chrongler_period} objects.
\item \code{<}: Compares the \code{start_date} of two \code{chrongler_period} objects.
\item \code{>}: Compares the \code{start_date} of two \code{chrongler_period} objects.
}
}
