% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exChronglerConc}
\alias{exChronglerConc}
\title{Example Concordance to be used with BuildingsMilet-Dataset}
\format{
An object of class \code{list} (inherits from \code{chrongler.conc}) of length 5.
}
\source{
\emph{L. Steinmann, S. Huy, F. Sliwka, D. Göçmen, N. Lordoğlu,
J. Zurbach, C. Berns (2023). Miletus Documentation Manual:
Surveying, Forms, Find Processing and Database-Usage, v2.0 (Version 2.0).
\url{http://doi.org/10.25592/uhhfdm.11389}}
}
\usage{
exChronglerConc
}
\description{
A subset of the periods and their translations used in the Miletus database.
Included to demonstrate and test the chrongler package with the \link{BuildingsMilet} dataset.
}
\seealso{
\link{BuildingsMilet}
}
\keyword{datasets}
