% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_periods.R
\name{group_periods}
\alias{group_periods}
\title{Replaces Periods with their Associated Grouped Versions}
\usage{
group_periods(data, conc, start, end)
}
\arguments{
\item{data}{\emph{data.frame}, \emph{required} containing at least two columns with the
start and end period for each row. Columns can be specified in the relevant
arguments. The values in the start and end period need to correspond with
the relevant values in the concordance used (see \code{\link[=make_chrongler_conc]{make_chrongler_conc()}}).}

\item{conc}{\emph{chrongler.conc}, \emph{required}, as built by \code{\link[=make_chrongler_conc]{make_chrongler_conc()}}.}

\item{start}{\emph{chr/int}, \emph{required}. Name or index of the column with the \strong{earliest period}.}

\item{end}{\emph{chr/int}, \emph{required}. Name or index of the column with the \strong{latest period}.}
}
\value{
The input \code{data} as a \code{data.frame}, with additional columns:
\itemize{
\item \code{start.grpd} -- \emph{ordered factor} of the group the value in the start column belongs to.
\item \code{end.grpd} -- \emph{ordered factor} of the group the value in the end column belongs to.
}
}
\description{
For each row in \code{data}, the group each period in the \code{start} and \code{end} column
belongs to is found. The groups are ordered factors as noted in the concordance
supplied to the \code{conc} (\code{\link[=make_chrongler_conc]{make_chrongler_conc()}}) argument. Values that
already are listed as groups are not changed.
}
\examples{
data("BuildingsMilet")
data("PeriodsMilet")
conc <- make_chrongler_conc(PeriodsMilet)
group_periods(
  BuildingsMilet,
  conc,
  start = "period.start",
  end = "period.end"
 )
}
\seealso{
\itemize{
\item \code{\link[=ungroup_periods]{ungroup_periods()}}
}
}
