% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrongler_period.R
\name{make_chrongler_periods}
\alias{make_chrongler_periods}
\title{Create Multiple chrongler_period Objects}
\usage{
make_chrongler_periods(
  name,
  start_date,
  end_date,
  group = NULL,
  color = NULL,
  source = NULL
)
}
\arguments{
\item{name}{A character string representing the name of the period.
This must be a single string. (Or: In \code{\link[=make_chrongler_periods]{make_chrongler_periods()}} a vector.)}

\item{start_date}{A numeric value representing the start date of the period.
It must be numeric and less than the end date. Negative values for dates BCE. (Or: In \code{\link[=make_chrongler_periods]{make_chrongler_periods()}} a vector.)}

\item{end_date}{A numeric value representing the end date of the period.
It must be numeric and greater than the start date. Negative values for dates BCE. (Or: In \code{\link[=make_chrongler_periods]{make_chrongler_periods()}} a vector.)}

\item{group}{(Optional) A character string representing the group to which
this period belongs. If not provided, the period name is used as the group. (Or: In \code{\link[=make_chrongler_periods]{make_chrongler_periods()}} a vector.)}

\item{color}{(Optional) A character string representing the color
associated with this period (e.g., a hex color code). (Or: In \code{\link[=make_chrongler_periods]{make_chrongler_periods()}} a vector.)}

\item{source}{(Optional) A character string representing the source
of the period's data or description. (Or: In \code{\link[=make_chrongler_periods]{make_chrongler_periods()}} a vector.)}
}
\value{
A list of \code{chrongler_period} objects, each representing a period
with the provided attributes. Invalid periods will be represented as \code{NA},
and warnings will be issued for any errors.
}
\description{
This function creates a list of \code{chrongler_period} objects from vectors
of names, start dates, end dates, and optional attributes such as group,
color, and source.
}
\details{
The function validates the inputs and ensures that all vectors are of
the same length. It utilizes the \code{\link[=new_chrongler_period]{new_chrongler_period()}} function to
generate each individual period.

The function ensures that all input vectors (name, start_date, end_date,
group, color, and source) have the same length. Duplicate names are not
allowed and will result in an error.

The actual creation of each period is handled by the
\code{\link[=new_chrongler_period]{new_chrongler_period()}} function, which is responsible for
validating and constructing each \code{chrongler_period} object.
}
\examples{
# Create multiple periods using vectors of attributes
names <- c("Early classical", "Late Classical", "Early hellenistic")
start_dates <- c(-480, -425, -323)
end_dates <- c(-426, -324, -178)
groups <- c("Classical", "Classical", "Hellenistic")
colors <- c("#582BA8", "#441794", "#283593")

periods <- make_chrongler_periods(names, start_dates, end_dates, groups, colors)

}
\seealso{
\link{new_chrongler_period} for the function used to
create individual \code{chrongler_period} objects.
}
