% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{na_if_empty}
\alias{na_if_empty}
\title{Return NA if a Value is NULL or of Length 0}
\usage{
na_if_empty(x)
}
\arguments{
\item{x}{A value.}
}
\value{
\code{x} or \code{NA}
}
\description{
Return NA if a Value is NULL or of Length 0
}
\examples{
na_if_empty(list())
na_if_empty(character(0))
na_if_empty(NULL)
}
\keyword{internal}
