% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{periods_exist}
\alias{periods_exist}
\title{Check if a value exists}
\usage{
periods_exist(period_vec, possible_values)
}
\arguments{
\item{period_vec}{A vector of values to check for}

\item{possible_values}{the vector against which to check}
}
\value{
TRUE if it exists, a character vector with non-existing values if not
}
\description{
Check if a value exists
}
\keyword{internal}
