% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PeriodsMilet}
\alias{PeriodsMilet}
\title{Periods as Listed in the Miletus Documentation Manual}
\format{
\subsection{\code{PeriodsMilet}}{

A data frame with 19 rows and 5 columns:
\describe{
\item{group}{The 'group' of periods the periods belongs to.}
\item{values}{The label of the period.}
\item{dating.min}{The "starting" year of the period.}
\item{dating.max}{The last year of the period.}
\item{color}{The color scheme as used by the Map of Miletus (see "BuildingsMilet" dataset).}
}
}
}
\source{
\url{https://doi.org/10.25592/uhhfdm.10029}
}
\usage{
PeriodsMilet
}
\description{
This is a slightly adapted set of periods as used by the Miletus Excavation,
for details see the "Miletus Documentation Manual":
Miletus Documentation Manual: Surveying, Forms, Find Processing and
Database Usage, v.2.0, \url{https://doi.org/10.25592/uhhfdm.10029}
}
\keyword{datasets}
