% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrongler_period.R
\name{print.chrongler_period}
\alias{print.chrongler_period}
\title{Print Method for chrongler_period}
\usage{
\method{print}{chrongler_period}(x, ...)
}
\arguments{
\item{x}{A \code{chrongler_period} object.}

\item{...}{Additional arguments passed to or from other methods.}
}
\description{
This function defines how \code{chrongler_period} objects are printed to the console.
It displays the name, start and end dates, group, color, and source (if available).
}
