% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{query_chronontology}
\alias{query_chronontology}
\title{(Removed) Basic Query for iDAI.chronontology}
\usage{
query_chronontology(value = "roman")
}
\arguments{
\item{value}{(chr)}
}
\value{
Nothing.
}
\description{
This function has been removed from chrongler.
}
