% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RomanPeriods}
\alias{RomanPeriods}
\title{Example Periods for the Roman Republic and the Roman Empire}
\format{
\subsection{\code{RomanPeriods}}{

A data frame with 8 rows and 6 columns:
\describe{
\item{group}{The 'group' of periods the periods belongs to.}
\item{values}{The label of the period.}
\item{dating.min}{The "starting" year of the period.}
\item{dating.max}{The last year of the period.}
\item{color}{A color for the period.}
\item{source}{A list of URL and comment on the source of the period.}
}
}
}
\source{
\url{https://chronontology.dainst.org}
}
\usage{
RomanPeriods
}
\description{
This is a set of Periods compiled from \href{https://chronontology.dainst.org}{iDAI.chronontology}
used to illustrate \code{chrongler}s functionality. It is compiled in the
vignette "chrongler wrangles categorical chronological data", see:
\code{vignette("chrongler_workflow", package = "chrongler")}
}
\keyword{datasets}
