% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ungroup_periods.R
\name{ungroup_periods}
\alias{ungroup_periods}
\title{Replaces Period-"groups" with their Associated Periods}
\usage{
ungroup_periods(data, conc, start, end)
}
\arguments{
\item{data}{\emph{data.frame}, \emph{required} containing at least two columns with the
start and end period for each row. Columns can be specified in the relevant
arguments. The values in the start and end period need to correspond with
the relevant values in the concordance used (see \code{\link[=make_chrongler_conc]{make_chrongler_conc()}}).}

\item{conc}{\emph{chrongler.conc}, \emph{required}, as built by \code{\link[=make_chrongler_conc]{make_chrongler_conc()}}.}

\item{start}{\emph{chr/int}, \emph{required}. Name or index of the column with the \strong{earliest period}.}

\item{end}{\emph{chr/int}, \emph{required}. Name or index of the column with the \strong{latest period}.}
}
\value{
The input \code{data} as a \code{data.frame}, with additional columns:
\itemize{
\item \code{start.ungr} -- \emph{ordered factor} of the earliest individual period in the group from the start column.
\item \code{end.ungr} -- \emph{ordered factor} of the latest individual period in the group from the end column.
}
}
\description{
For each row in \code{data}, if the value in the \code{start} and \code{end} column
is listed as a \emph{group} in the concordance object supplied to
\code{conc} (\code{\link[=make_chrongler_conc]{make_chrongler_conc()}}), it is replaced with the name of the
first or last respective \emph{period} in that \emph{group} as noted in said concordance.
Values that are already listed as \emph{period}s are not changed.
}
\examples{
data("BuildingsMilet")
data("PeriodsMilet")
conc <- make_chrongler_conc(PeriodsMilet)
ungroup_periods(
  BuildingsMilet,
  conc,
  start = "period.start",
  end = "period.end"
 )
}
\seealso{
\itemize{
\item \code{\link[=group_periods]{group_periods()}}
}
}
