% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_dating.R
\name{derive_dating}
\alias{derive_dating}
\title{Derive Dating from Periods}
\usage{
derive_dating(data, conc, start, end, dating.min, dating.max)
}
\arguments{
\item{data}{A data.frame containing at least two columns with the
start and end period for each row. Columns can be specified in the relevant
arguments. The values in the start and end period need to correspond with
the relevant values in the correspondence object used
(see \code{\link[=make_chrongler_conc]{make_chrongler_conc()}}).}

\item{conc}{A concordance list build by \code{\link[=make_chrongler_conc]{make_chrongler_conc()}}.}

\item{start}{\emph{chr/int}. Name or index of the column with the \strong{minimum period}.}

\item{end}{\emph{chr/int}. Name or index of the column with the \strong{maximum period}.}

\item{dating.min}{\emph{chr/int}. Name or index of a pre-existing column
containing absolute \strong{minimum dating}. If it does exist, this function will
only fill cells that were previously empty. A comment is stored in a
column called \emph{dating.source}. If this argument is left empty and a \emph{dating.min}
column already exists, it will be overwritten.}

\item{dating.max}{\emph{chr/int}. Name or index of a pre-existing column
containing absolute \strong{maximum dating}. If it does exist, this function will
only fill cells that were previously empty. A comment is stored in a
column called \emph{dating.source}. If this argument is left empty and a \emph{dating.max}
column already exists, it will be overwritten.}
}
\value{
A data.frame, with absolute values in the \code{dating.min} and
\code{dating.max} columns and a comment in \code{dating.source} (see description).
}
\description{
For each row in \code{data}, the value in the \code{start} and \code{end} column
is used to gather the absolute dating of its period from the concordance
object supplied to \code{conc} (\code{\link[=make_chrongler_conc]{make_chrongler_conc()}}). The absolute dating
is stored in the \code{dating.min} and \code{dating.max} columns, which have this
name if the corresponding arguments are empty. If column names are
supplied as arguments, the data in these columns will not be overwritten
and instead only filled in if the respective cells are empty (NA).
A comment informing of the procedure is written to a column called
\emph{dating.source}.
}
\examples{
data("BuildingsMilet")

filename <- system.file(package = "chrongler",
                       "extdata/2023_periods_grouping_example.csv")
conc <- make_chrongler_conc(filename)

derive_dating(BuildingsMilet, conc, start = "period.start", end = "period.end")
}
\seealso{
\code{\link[=make_chrongler_conc]{make_chrongler_conc()}}
}
