% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_period.R
\name{derive_period}
\alias{derive_period}
\title{Derive Period from Absolute Dates}
\usage{
derive_period(
  data,
  conc,
  dating.min,
  dating.max,
  start,
  end,
  by_group = FALSE,
  paste_multiple = TRUE
)
}
\arguments{
\item{data}{A data.frame containing at least two columns with the
start and end period for each row. Columns can be specified in the relevant
arguments. The values in the start and end period need to correspond with
the relevant values in the correspondence object used
(see \code{\link[=make_chrongler_conc]{make_chrongler_conc()}}).}

\item{conc}{A concordance list build by \code{\link[=make_chrongler_conc]{make_chrongler_conc()}}.}

\item{dating.min}{chr/int}

\item{dating.max}{chr/int}

\item{start}{\emph{chr/int}. Name or index of a pre-existing column
containing the \strong{starting period}. If it does exist, this function will
only fill cells that were previously empty. A comment is stored in a
column called \emph{period.source}. If this argument is left empty and a
\emph{period.start} column already exists, it will be overwritten.}

\item{end}{\emph{chr/int}. Name or index of a pre-existing column
containing the \strong{ending period}. If it does exist, this function will
only fill cells that were previously empty. A comment is stored in a
column called \emph{period.source}. If this argument is left empty and a
\emph{period.end} column already exists, it will be overwritten.}

\item{by_group}{TRUE/FALSE}

\item{paste_multiple}{TRUE/FALSE}
}
\value{
A data.frame, with period values in the \code{period.start} and
\code{period.end} columns and a comment in \code{period.source} (see description).
}
\description{
For each row in \code{data}, the value in the \code{dating.min} and \code{dating.max} column
is used to identify the period(s) the respective row would be dated in
according to the dating ranges registered in the the concordance
object supplied to \code{conc} (\code{\link[=make_chrongler_conc]{make_chrongler_conc()}}). The periods are
stored in the \code{period.start} and \code{period.end} columns, which have this
name if the corresponding arguments are empty. If column names are
supplied as arguments, the data in these columns will not be overwritten
and instead only filled in if the respective cells are empty (NA).
A comment informing of the procedure is written to a column called
\emph{period.source}.
}
\examples{
filename <- system.file(package = "chrongler",
                       "extdata/2023_periods_grouping_example.csv")
conc <- make_chrongler_conc(filename)

data <- data.frame(name = 1:110)

data$dating.min <- sample(seq(-470, 1999, by = 1), 110, replace = TRUE)
data$dating.max <- data$dating.min + sample(round(rnorm(1000, 30, 10)), 110, replace = TRUE)

derive_period(data, conc,
              dating.min = "dating.min",
              dating.max = "dating.max",
              by_group = FALSE, paste_multiple = FALSE)

}
\seealso{
\itemize{
\item \code{\link[=make_chrongler_conc]{make_chrongler_conc()}}
\item \code{\link[=derive_dating]{derive_dating()}}
}
}
