% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_by.R
\name{duplicate_by}
\alias{duplicate_by}
\title{Duplicate each Row for each Period in the Range from start to end}
\usage{
duplicate_by(data, conc, start, end, by_group = TRUE)
}
\arguments{
\item{data}{A data.frame containing at least two columns with the
start and end period for each row. Columns can be specified in the relevant
arguments. The values in the start and end period need to correspond with
the relevant values in the correspondence object used
(see \code{\link[=make_chrongler_conc]{make_chrongler_conc()}}).}

\item{conc}{A concordance list build by \code{\link[=make_chrongler_conc]{make_chrongler_conc()}}.}

\item{start}{\emph{chr/int}. Name or index of the column with the \strong{minimum period}.}

\item{end}{\emph{chr/int}. Name or index of the column with the \strong{maximum period}.}

\item{by_group}{\emph{TRUE/FALSE}. Default is TRUE. If TRUE all rows are
duplicated according to their grouped periods (see \code{\link[=group_periods]{group_periods()}}). If
FALSE all rows are duplicated according to their single
periods (see \code{\link[=ungroup_periods]{ungroup_periods()}}).}
}
\value{
A data.frame, with a row for each period in which each row from the
previous data would fall (see description).
}
\description{
Duplicate each Row for each Period in the Range from start to end
}
\examples{
data("BuildingsMilet")

filename <- system.file(package = "chrongler",
                       "extdata/2023_periods_grouping_example.csv")
conc <- make_chrongler_conc(filename)

duplicate_by(BuildingsMilet, conc,
             start = "period.start", end = "period.end",
             by_group = FALSE)

}
