% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{from_chronontology}
\alias{from_chronontology}
\title{(Removed) Get Data for One Period from iDAI.chronontology
This function has been removed from chrongler.}
\usage{
from_chronontology(id = "")
}
\arguments{
\item{id}{(chr)}
}
\value{
Nothing.
}
\description{
(Removed) Get Data for One Period from iDAI.chronontology
This function has been removed from chrongler.
}
