% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_chronontology.R
\name{from_chronontology}
\alias{from_chronontology}
\title{Get Data for One Period from iDAI.chronOntology}
\source{
data: \href{https://chronontology.dainst.org/}{iDAI.chronontology}
}
\usage{
from_chronontology(id = "")
}
\arguments{
\item{id}{(chr) The chronOntology-ID of the period that should be
downloaded (part of the URL after
"https://chronontology.dainst.org/period/period/").}
}
\value{
A nested list with the data for one period downloaded
from iDAI.chronOntology.
}
\description{
Get Data for One Period from iDAI.chronOntology
}
\examples{
geometric <- from_chronontology("AqUgeWHp0xIl")
}
