% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_periods.R
\name{group_periods}
\alias{group_periods}
\title{Replaces Periods with their Associated Grouped Versions}
\usage{
group_periods(data, conc, start, end)
}
\arguments{
\item{data}{A data.frame containing at least two columns with the
start and end period for each row. Columns can be specified in the relevant
arguments. The values in the start and end period need to correspond with
the relevant values in the correspondence object used
(see \code{\link[=make_chrongler_conc]{make_chrongler_conc()}}).}

\item{conc}{A concordance list build by \code{\link[=make_chrongler_conc]{make_chrongler_conc()}}.}

\item{start}{\emph{chr/int}. Name or index of the column with the \strong{minimum period}.}

\item{end}{\emph{chr/int}. Name or index of the column with the \strong{maximum period}.}
}
\value{
A data.frame, with the values replaced in the columns from
the \code{start} and \code{end}-arguments (see description).
}
\description{
For each row in \code{data}, the value in the \code{start} and \code{end} column is
replaced with the name of the respective \emph{group} as noted in the concordance
supplied to the \code{conc} (\code{\link[=make_chrongler_conc]{make_chrongler_conc()}}) argument. Values that
already are listed as groups are not changed.
}
\examples{
data("BuildingsMilet")
conc <- read.csv(system.file(package = "chrongler",
  "extdata/2023_periods_grouping_example.csv"))
conc <- make_chrongler_conc(conc)
group_periods(
  BuildingsMilet,
  conc,
  start = "period.start",
  end = "period.end"
 )
}
\seealso{
\code{\link[=make_chrongler_conc]{make_chrongler_conc()}}
}
