% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_chrongler_conc.R
\name{make_chrongler_conc}
\alias{make_chrongler_conc}
\title{Produce a concordance object that can be used with \code{chrongler} functions}
\usage{
make_chrongler_conc(
  file,
  cols = list(group = NA, values = NA, dating.min = NA, dating.max = NA, color = NA),
  ...
)
}
\arguments{
\item{file}{chr / data.frame / matrix}

\item{cols}{list with specific values and column names or Indices, see description.}
}
\value{
A list.
}
\description{
Produce a concordance object that can be used with \code{chrongler} functions
}
\details{
If the column names of your provided csv file or data.frame already
conform to the following logic, you can omit the \code{cols}-argument altogether.
If not, you need to provide the a list in the following format to \code{cols}:
\itemize{
\item \strong{groups} (\emph{chr/int}) Name or Index of the column that contains
the grouping variable.
\item \strong{values} (\emph{chr/int}) Name or Index of the column that contains
all period names. These are expected to be in chronological order!
\item \strong{dating.min} (\emph{chr/int}) Name or Index of the column that contains
the minimum / 'from' absolute date for each period.
Expects negative values for BCE and positive values for CE.
\item \strong{dating.max} (\emph{chr/int}) Name or Index of the column that contains
the maximum / 'to' absolute date for each period.
Expects negative values for BCE and positive values for CE.
\item \strong{color} (\emph{chr/int}) Name or Index of the column that contains
the color associated with each value for consistent
color scales in plots.
}
}
\examples{
\dontrun{
file <- system.file(package = "chrongler",
                    "extdata/2023_periods_grouping_example.csv")
make_chrongler_conc(file)
}
}
