% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{missing_periods}
\alias{missing_periods}
\title{Check for and Return Periods not Present in the Concordance}
\usage{
missing_periods(period_vec, possible_periods)
}
\arguments{
\item{period_vec}{A vector of values to check for}

\item{possible_periods}{The vector of values against which to check}
}
\value{
a character vector with the values from \code{period_vec} not present
in \code{possible_periods}, or \code{character(0)} if all are present.
}
\description{
Check for and Return Periods not Present in the Concordance
}
\keyword{internal}
