% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_chronontology.R
\name{query_chronontology}
\alias{query_chronontology}
\title{Basic Query for iDAI.chronontology}
\usage{
query_chronontology(value = "roman")
}
\arguments{
\item{value}{(chr) The value you want to search for in all chronontology-periods.}
}
\value{
A list of results as handed out by iDAI.chronontology.
}
\description{
Basic Query for iDAI.chronontology
}
\examples{
roman_periods <- query_chronontology(value = "roman")
names <- unlist(lapply(roman_periods, function(x) x$resource$names$en))
}
