% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ungroup_periods.R
\name{ungroup_periods}
\alias{ungroup_periods}
\title{Replaces Period-"groups" with their Associated Periods}
\usage{
ungroup_periods(data, conc, start, end)
}
\arguments{
\item{data}{A data.frame containing at least two columns with the
start and end period for each row. Columns can be specified in the relevant
arguments. The values in the start and end period need to correspond with
the relevant values in the correspondence object used
(see \code{\link[=make_chrongler_conc]{make_chrongler_conc()}}).}

\item{conc}{A concordance list build by \code{\link[=make_chrongler_conc]{make_chrongler_conc()}}.}

\item{start}{\emph{chr/int}. Name or index of the column with the \strong{minimum period}.}

\item{end}{\emph{chr/int}. Name or index of the column with the \strong{maximum period}.}
}
\value{
A data.frame, with the values replaced in the columns from
the \code{start} and \code{end}-arguments (see description).
}
\description{
For each row in \code{data}, if the value in the \code{start} and \code{end} column
is listed as a \emph{group} in the concordance object supplied to
\code{conc} (\code{\link[=make_chrongler_conc]{make_chrongler_conc()}}), it is replaced with the name of the
first respective \emph{period} in that \emph{group} as noted in said concordance.
Values that are already listed as \emph{period}s are not changed.
}
\examples{
data("BuildingsMilet")
filename <- system.file(package = "chrongler",
                       "extdata/2023_periods_grouping_example.csv")
conc <- make_chrongler_conc(filename)
ungroup_periods(
  BuildingsMilet,
  conc,
  start = "period.start",
  end = "period.end"
 )
}
\seealso{
\code{\link[=make_chrongler_conc]{make_chrongler_conc()}}
}
